Change BaseArray class so it doesn't automatically convert to floats
Now an array_like of integers will have a numpy int dtype.

+++ skspatial/objects/_base_array.py
@@ -8,7 +8,7 @@ class _BaseArray(np.ndarray):
 
     def __new__(cls, array_like):
 
-        array = np.array(array_like, dtype=float)
+        array = np.array(array_like)
 
         if array.size == 0:
             raise ValueError("The array must not be empty.")
